Replaced the `sed` command in build_typescript.sh.
The `sed` command as written will only work on Linux/GNU variants
of `sed`, and is incompatile with BSD / OSX variants of sed.

I was able to modify two of the three `sed` invocations to work
cross-platform, but the third (and most complex) `sed` command
I could not get to work - so instead I replaced it with a small
node.js script which both copies the files and does the textual
modifications.

The script has no dependencies other than node.js, and given that
the build script requires typescript to be installed, presumably
the user will have node available.