Separate generated JavaScript from TypeScript source.
Changes the typescript_build.sh script to copy the .ts
files into 'pkg/src' instead of 'pkg'. Adjusts the tsconfig.json
file so that it looks for sources in both 'pkg' and 'pkg/src'.

Verified that generated source maps point to the correct location.

In general, mixing .ts sources and generated .js and .d.ts in the
same directory can cause problems for downstream consumers of
the library. Some bundlers will preferentially load the .ts files
instead of the .js files, which will generate syntax errors if
the TypeScript compiler options for the main app are significantly
different than the compiler options for the library. By keeping
these files separate, better isolation is achieved.