Overhaul of the rapier-compat build.
The new build is based on rollup rather than webpack. Common wisdom
is that Rollup is better for libraries. There are a number of concrete
advantages:

* The build is signficantly faster.
* The minified JavaScript is considerably smaller.
* There is a lot less complex boilerplate code in the bundle.
* And most importantly, the build now generates both an ES module
  and a CommonJS module. These are both exported from package.json,
  so bundlers can select whichever one they like.

ES module support is significant because it allows tree-shaking
optimization, so applications don't have to include parts of the library
they are not using. That being said, there are two caveats:

1) Most of the bundle size is the WASM blob, which isn't affected
   by tree-shaking.
2) Tree-shaking only works if you are importing individual symbols:

   ```ts
   import { init, Capsule } from '@dimforge/rapier3d-compat';
   ```

   If you use the default import, then you get everything, because
   it's an object:

   ```ts
   import RAPIER from '@dimforge/rapier3d-compat';
   ```

This commit also eliminates the 'two-pass' webpack build. Rollup is
only invoked once. The generation of the 2d/3d sources are performed
by a shell script similar to the one used in rapier2d/rapier3d.

Some of the more complex steps of the webpack build - such as generating
the `init` function - have been replaced by separate source directories.
The TypeScript 'rootDirs' config option is used to select between the
2d and 3d init functions. (The 'init' function is also considerably
simpler, thanks to Rollup's base64 plugin.)

The build no longer copies the source tree into the output directory,
but instead relies on Rollup to generate a proper sourcemap file.

In addition, this commit also includes a number of minor TypeScript
fixes, as well as tightening up the tsconfig options a bit. (However,
the most important strict mode, 'strictNullChecks' is still disabled.)

Also, removed some circular dependencies that Rollup was complaining
about. One circular dependency remains, but it is a necessary one.

This change should be completely backwards compatible.